backend/requirements.txt
fastapi
uvicorn
python-socketio
redis
upstash-redis
python-multipart
requests
gradio_client
